# src/swagger.json
{
	"openapi": "3.0.0",
	"info": {
	  "title": "Comeback",
	  "contact": {},
	  "version": "1.0"
	},
	"servers": [
	  {
		"url": "https://comeback-backend.herokuapp.com/",
		"variables": {}
	  }
	],
	"paths": {
...
	  "/feedback": {
		"get": {
		  "tags": [
			"Feedback"
		  ],
		  "summary": "Get All",
		  "operationId": "GetAll1234",
		  "parameters": [
			{
			  "name": "Content-Type",
			  "in": "header",
			  "description": "",
			  "required": true,
			  "style": "simple",
			  "schema": {
				"type": "string",
				"example": "application/json"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "",
			  "headers": {}
			}
		  },
		  "deprecated": false
		},
		"put": {
		  "tags": [
			"Feedback"
		  ],
		  "summary": "Update",
		  "operationId": "PutUpdate1234",
		  "parameters": [],
		  "requestBody": {
			"description": "",
			"content": {
			  "application/json": {
				"schema": {
				  "$ref": "#/components/schemas/UpdateRequest"
				},
				"example": {
				  "id": "603071953f3d4f418cbdb5b4",
				  "data": {}
				}
			  }
			},
			"required": true
		  },
		  "responses": {
			"200": {
			  "description": "",
			  "headers": {}
			}
		  },
		  "deprecated": false
		},
		"delete": {
		  "tags": [
			"Feedback"
		  ],
		  "summary": "Delete",
		  "operationId": "Delete12345",
		  "parameters": [],
		  "requestBody": {
			"description": "",
			"content": {
			  "application/json": {
				"schema": {
				  "$ref": "#/components/schemas/DeleteRequest"
				},
				"example": {
				  "id": "602f68e7dbfc3a1bb43ad1c3"
				}
			  }
			},
			"required": true
		  },
		  "responses": {
			"200": {
			  "description": "",
			  "headers": {}
			}
		  },
		  "deprecated": false
		}
	  },
...
	  "/appoiments": {
		"post": {
		  "tags": [
			"Appoiments"
		  ],
		  "summary": "Add",
		  "operationId": "PostAdd123",
		  "parameters": [],
		  "requestBody": {
			"description": "",
			"content": {
			  "application/json": {
				"schema": {
				  "$ref": "#/components/schemas/AddRequest4"
				},
				"example": {
				  "name": "asdasdasd",
				  "hours": "asdasdas",
				  "professionalID": "asdasdas",
				  "userID": "asdasdas",
				  "day": "asdasdas",
				  "month": "asdasdas",
				  "year": "asdasdas"
				}
			  }
			},
			"required": true
		  },
		  "responses": {
			"200": {
			  "description": "",
			  "headers": {}
			}
		  },
		  "deprecated": false
		},
		"get": {
		  "tags": [
			"Appoiments"
		  ],
		  "summary": "Get All",
		  "operationId": "GetAll123456",
		  "parameters": [
			{
			  "name": "Content-Type",
			  "in": "header",
			  "description": "",
			  "required": true,
			  "style": "simple",
			  "schema": {
				"type": "string",
				"example": "application/json"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "",
			  "headers": {}
			}
		  },
		  "deprecated": false
		},
		"put": {
		  "tags": [
			"Appoiments"
		  ],
		  "summary": "Update",
		  "operationId": "PutUpdate123456",
		  "parameters": [],
		  "requestBody": {
			"description": "",
			"content": {
			  "application/json": {
				"schema": {
				  "$ref": "#/components/schemas/UpdateRequest"
				},
				"example": {
				  "id": "603071953f3d4f418cbdb5b4",
				  "data": {}
				}
			  }
			},
			"required": true
		  },
		  "responses": {
			"200": {
			  "description": "",
			  "headers": {}
			}
		  },
		  "deprecated": false
		},
		"delete": {
		  "tags": [
			"Appoiments"
		  ],
		  "summary": "Delete",
		  "operationId": "Delete1234567",
		  "parameters": [],
		  "requestBody": {
			"description": "",
			"content": {
			  "application/json": {
				"schema": {
				  "$ref": "#/components/schemas/DeleteRequest"
				},
				"example": {
				  "id": "602f68e7dbfc3a1bb43ad1c3"
				}
			  }
			},
			"required": true
		  },
		  "responses": {
			"200": {
			  "description": "",
			  "headers": {}
			}
		  },
		  "deprecated": false
		}
	  },
...
	},
	"components": {
	  "schemas": {
...
		"UpdateRequest": {
		  "title": "UpdateRequest",
		  "required": [
			"id",
			"data"
		  ],
		  "type": "object",
		  "properties": {
			"id": {
			  "type": "string"
			},
			"data": {
			  "type": "object"
			}
		  },
		  "example": {
			"id": "602f68e7dbfc3a1bb43ad1c3",
			"data": {}
		  }
		},
		"DeleteRequest": {
		  "title": "DeleteRequest",
		  "required": [
			"id"
		  ],
		  "type": "object",
		  "properties": {
			"id": {
			  "type": "string"
			}
		  },
		  "example": {
			"id": "602f68e7dbfc3a1bb43ad1c3"
		  }
		},
...
		"AddRequest4": {
		  "title": "AddRequest4",
		  "required": [
			"name",
			"hours",
			"professionalID",
			"userID",
			"day",
			"month",
			"year"
		  ],
		  "type": "object",
		  "properties": {
			"name": {
			  "type": "string"
			},
			"hours": {
			  "type": "string"
			},
			"professionalID": {
			  "type": "string"
			},
			"userID": {
			  "type": "string"
			},
			"day": {
			  "type": "string"
			},
			"month": {
			  "type": "string"
			},
			"year": {
			  "type": "string"
			}
		  },
		  "example": {
			"name": "asdasdasd",
			"hours": "asdasdas",
			"professionalID": "asdasdas",
			"userID": "asdasdas",
			"day": "asdasdas",
			"month": "asdasdas",
			"year": "asdasdas"
		  }
		},
...
	  },
	  "securitySchemes": {
		"httpBearer": {
		  "type": "http",
		  "scheme": "bearer"
		}
	  }
	},
	"security": [
	  {
		"httpBearer": []
	  }
	],
...
  }